Moved pack contents to inventory

+++ build.ps1
@@ -1,11 +1,16 @@
+Param(
+  [switch]$skipBuild = $false
+)
 $ErrorActionPreference = "Stop"
 
 yaml2json -s "m-chee-daro.yml"
 
 cd app
-npm install
-npm run build
-Remove-Item -Recurse -Path ../docs -Exclude "_config.yml" -Force
+if ($skipBuild -eq $false){
+  npm install
+  npm run build
+  Remove-Item -Recurse -Path ../docs -Exclude "_config.yml" -Force
+}
 
 Copy-Item -Force "../m-chee-daro.json" build
 Copy-Item -Force "../m-chee-daro.json" src
